Fixed position constraint
Along with more work for the rigidbody system

+++ Hazel/Source/Hazel/Physics/SystemState.h
@@ -11,8 +11,8 @@ namespace Enyoo
 		SystemState();
 		~SystemState();
 
-		glm::vec2 LocalToWorld(glm::dvec2 point, int index);
-		glm::vec2 VelocityAtPoint(glm::dvec2 point, int index);
+		glm::dvec2 LocalToWorld(glm::dvec2 point, int index);
+		glm::dvec2 VelocityAtPoint(glm::dvec2 point, int index);
 		void ApplyForce(glm::dvec2 point, glm::dvec2 force, int index);
 		void Resize(uint32_t count);
 		void Destroy();
